backend/requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
